refactor: moved SetLenOnDrop to set_len_on_drop
